requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
seqlog==0.3.30
six==1.16.0
urllib3==2.0.7
